adjusted Rewriterule to remove request URI
The rule I was using would include the request URI. This is ideal to redirect
something like http://example.com/my/path to https://example.com/my/path but
it was dropping the requester on the directory index for /var/www rather than
/var/www/html where the html index file resides. It's unclear to me why this
was happening. This should be a sufficient workaround for the time being.